+++ src/app/view/WelcomeComponent.js
@@ -122,9 +122,13 @@ class WelcomeComponent extends Component {
 					</div>
 					
 					<Typography variant="subheading">{I18n.t("Our free-licensed pictures come from various communities")}</Typography>
-					{Object.entries((new P4C.PicturesManager()).getFetcherDetails()).map(e => (
-						<a href={e[1].homepageUrl} key={e[0]}><img src={e[1].logoUrl} title={e[1].name} alt={e[1].name} style={{ maxHeight: 48, maxWidth: 48, margin: "10px 20px", verticalAlign: "middle" }} /></a>
-					))}
+					{Object.entries((new P4C.PicturesManager())
+						.getFetcherDetails())
+						.filter(e => e[0] !== "flickr")
+						.map(e => (
+							<a href={e[1].homepageUrl} key={e[0]}><img src={e[1].logoUrl} title={e[1].name} alt={e[1].name} style={{ maxHeight: 48, maxWidth: 48, margin: "10px 20px", verticalAlign: "middle" }} /></a>
+						))
+					}
 				</div>
 				
 				<Grid container justify="center">